pyproject.toml
[project]
name = "generative-models-api"
version = "1.0.0"
description = "API for generating images using GAN, Energy-Based Models, and Diffusion Models"
readme = "README.md"
requires-python = ">=3.9"
authors = [
    {name = "Michael Yin"}
]

dependencies = [
    "torch>=2.0.0",
    "torchvision>=0.15.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.0.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    "tqdm>=4.65.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "httpx>=0.25.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["app", "app.helper_lib"]

[tool.black]
line-length = 100
target-version = ['py39']

[tool.ruff]
line-length = 100
target-version = "py39"

